+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New js helpers in `jarboe` object: `bigToastSuccess`, `bigToastDanger`, `bigToastWarning`, `bigToastInfo`.
 - Inline editing for `Checkbox` field.
 - Translation keys `jarboe::fields.checkbox.yes` and `jarboe::fields.checkbox.no` that are used in `Checkbox` field inline mode.
+- `FieldsetMarkup` field for dividing form into blocks with its own `<legend>`.
 
 ## [1.1.0] - 2020-06-16
 ### Added
@@ -28,5 +29,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 [Unreleased]: https://github.com/Cherry-Pie/Jarboe/compare/1.1.0...master
-[1.0.0]: https://github.com/Cherry-Pie/Jarboe
 [1.1.0]: https://github.com/Cherry-Pie/Jarboe/compare/1.0.0...1.1.0
+[1.0.0]: https://github.com/Cherry-Pie/Jarboe